Use alpine-node 6.2.0 instead of 6.2
With latest update in the alpine-node:6.2 (to version/tag 6.2.1)
something has changed and other builds are failing.

 It seems that SSL is
not enabled by default (or installed) in the latest version and wget and
others are failing.

Also, we cannot overwrite the current node version with an older one for
some components that require it (and in 6.2 we could do that).